VisIsuNi-Restitution/functions/otPereDefautsMaj/requirements.txt
functions-framework
google-cloud-bigquery
Flask